// src/lib/types.ts
export interface Therapist {
  id: number
  name: string
  speciality: string[]
  location: string
  availability: string[]
  rating: number
  description: string
  qualification: string
  experience: number
  format: string[]
  gender: string
  languages: string[]
  profileImage: string
  coverImage: string
}